feat: [factory-argument-signatures-prd-mergeability] - [Restore package-count mergeability]
